Added an enemy
Added in another capsule as enemy. The enemy has a navmesh agent attached to it. So it can chase around the player. Bullets can now collide with enemy. Enemy's capsule collider is set to trigger so that it can't throw the player off the ground or push him around. To improve performance navmesh agent updates only four times a second.